Refactor animation
- Change AnimationDict (map) to AnimationSet (a full class) which track currentAnimation using Play()
 - AnimationSet
  - currentAnimation
  - map<name, Animation> animations
  - Play()
    - currentAnimation = newAnimation
- Animation
  - isLooping = true default
  - IsComplete() = true => last frame && isLooping = false

+++ src/GameCuaTao/Direct2DGame/Animations/AnimationFrame.cpp
@@ -4,13 +4,13 @@ AnimationFrame::AnimationFrame()
 {
 }
 
-AnimationFrame::AnimationFrame(Sprite sprite, unsigned long time)
+AnimationFrame::AnimationFrame(Sprite sprite, int time)
 {
 	this->sprite = sprite;
 	this->time = time;
 }
 
-unsigned long AnimationFrame::GetTime()
+int AnimationFrame::GetTime()
 {
 	return time;
 }
